fileupload/requirements.txt
Django==1.11.20
djangorestframework==3.11.1
pytz==2020.1
